% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charisma2.R
\name{charisma2}
\alias{charisma2}
\title{Add together two numbers}
\usage{
charisma2(
  charisma.obj,
  interactive = T,
  new.threshold = NULL,
  which.state = c("none", "merge", "replace"),
  state.index = NULL
)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
