% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charisma_to_img.R
\name{charisma_to_img}
\alias{charisma_to_img}
\title{Add together two numbers}
\usage{
charisma_to_img(
  charisma_obj,
  out_type = c("jpg", "jpeg", "png"),
  bg_color = "white",
  render_method = c("array", "raster"),
  filename = ""
)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
