% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color2label.R
\name{color2label}
\alias{color2label}
\title{Add together two numbers}
\usage{
color2label(color_triplet, hsv = F, verbose = F, clut = charisma::clut)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
