% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mapped_hex.R
\name{get_mapped_hex}
\alias{get_mapped_hex}
\title{Add together two numbers}
\usage{
get_mapped_hex(mapping = color.map)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
