% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCentroidDistances.R
\name{getCentroidDistances}
\alias{getCentroidDistances}
\title{Add together two numbers}
\usage{
getCentroidDistances(charisma_obj, mapping = color.map)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
