% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColor.R
\name{getColor}
\alias{getColor}
\title{Add together two numbers}
\usage{
getColor(
  color_triplet,
  color.space = c("rgb", "hsv"),
  verbose = FALSE,
  mapping = color.map
)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
