% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatchedHex.R
\name{getMatchedHex}
\alias{getMatchedHex}
\title{Add together two numbers}
\usage{
getMatchedHex(color_name, hex_vector)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
