% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialDensityScores.R
\name{getSpatialDensityScores}
\alias{getSpatialDensityScores}
\title{Add together two numbers}
\usage{
getSpatialDensityScores(charisma_obj, mapping = charisma::color.map)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
