% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummary.R
\name{getSummary}
\alias{getSummary}
\title{Add together two numbers}
\usage{
getSummary(
  charisma_obj,
  method = c("both", "freq", "spatial"),
  freq.threshold = 0.5,
  spatial.threshold = 0.5,
  mapping = color.map
)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
