% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_image.R
\name{load_image}
\alias{load_image}
\title{Add together two numbers}
\usage{
load_image(
  img_path = system.file("extdata/corbetti.png", package = "recolorize"),
  verbose = TRUE,
  plot = TRUE
)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
