% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pavo_classify_charisma.R
\name{pavo_classify_charisma}
\alias{pavo_classify_charisma}
\title{Add together two numbers}
\usage{
pavo_classify_charisma(charisma_obj, plot = TRUE)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
