% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_charisma.R
\name{plot.charisma}
\alias{plot.charisma}
\title{this is a test}
\usage{
\method{plot}{charisma}(x, ...)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
