% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagnostic.R
\name{plotDiagnostic}
\alias{plotDiagnostic}
\title{Add together two numbers}
\usage{
plotDiagnostic(
  charisma_obj,
  mapping = charisma::color.map,
  freq.threshold = 0.05,
  spatial.threshold = 0.01
)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
