% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPipeline.R
\name{runPipeline}
\alias{runPipeline}
\title{Add together two numbers}
\usage{
runPipeline(
  images,
  mapping = charisma::color.map,
  summary.method = c("both", "freq", "spatial"),
  freq.threshold = 0.05,
  spatial.threshold = 0.05,
  validate.color.map = FALSE,
  validation.simple = TRUE,
  lower = NULL,
  upper = NULL,
  alpha.channel = TRUE,
  plot.diagnostic = FALSE,
  save.plots = FALSE,
  output.dir = getwd(),
  save.plot.type = c("pdf", "png", "jpeg", "tiff", "bmp"),
  plot.width = 10,
  plot.height = 5
)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
