% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateColorMap.R
\name{validateColorMap}
\alias{validateColorMap}
\title{Add together two numbers}
\usage{
validateColorMap(mapping = charisma::color.map, simple = TRUE)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
