% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charisma.R
\name{charisma}
\alias{charisma}
\title{Add together two numbers}
\usage{
charisma(
  img_path,
  threshold = 0,
  auto.drop = T,
  interactive = F,
  plot = F,
  pavo = T,
  logdir = NULL,
  stack_colors = T,
  bins = 4,
  cutoff = 20,
  clut = charisma::clut
)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
