% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_charisma.R
\name{plot.charisma}
\alias{plot.charisma}
\title{this is a test}
\usage{
\method{plot}{charisma}(
  x,
  plot.all = T,
  plot.original = F,
  plot.recolored = F,
  plot.masked = F,
  plot.props = F,
  plot.pavo.img = F,
  plot.pavo.classes = F,
  font.size = 1.75,
  real.bar.colors = T
)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
