% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.charisma.R
\name{plot.charisma}
\alias{plot.charisma}
\title{Add together two numbers}
\usage{
\method{plot}{charisma}(
  x,
  ...,
  plot.original = TRUE,
  plot.sprite = TRUE,
  plot.spatial = TRUE,
  plot.centroids = FALSE,
  plot.centroids.type = c("alpha", "silhouette", "original"),
  freq.threshold = 0.5,
  spatial.threshold = 0.5,
  centroid.threshold = 0.5
)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
