% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotColors.R
\name{plotColors}
\alias{plotColors}
\title{Add together two numbers}
\usage{
plotColors(
  charisma_obj,
  type = c("freq", "spatial"),
  threshold = 0.05,
  multi.plot = FALSE,
  mapping = color.map
)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
