% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSprite.R
\name{plotSprite}
\alias{plotSprite}
\title{Add together two numbers}
\usage{
plotSprite(
  charisma_obj,
  centroids = FALSE,
  plot.centroids.type = c("alpha", "silhouette", "original"),
  cex = 3,
  multi.plot = FALSE,
  mapping = color.map
)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)

}
